Extend queuing to character console
